pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "gtts>=2.5.4",
    "openai>=1.107.1",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.49.1",
]
